Move ItemTable to Tables module
