+++ src/Engine/Events/EventInterpreter.cpp
@@ -7,14 +7,14 @@
 #include "Engine/Events/EventIR.h"
 #include "Engine/Events/Processor.h"
 #include "Engine/Party.h"
+#include "Engine/LOD.h"
 #include "Engine/Graphics/IRender.h"
 #include "Engine/Graphics/Indoor.h"
 #include "Engine/Graphics/Weather.h"
 #include "Engine/Graphics/Level/Decoration.h"
 #include "Engine/Objects/NPC.h"
 #include "Engine/Objects/SpriteObject.h"
-#include "Engine/Objects/ItemTable.h"
-#include "Engine/LOD.h"
+#include "Engine/Tables/ItemTable.h"
 
 #include "Media/Audio/AudioPlayer.h"
 #include "Media/MediaPlayer.h"
