+++ src/Engine/Graphics/Outdoor.cpp
@@ -22,12 +22,12 @@
 #include "Engine/LOD.h"
 #include "Engine/Objects/Actor.h"
 #include "Engine/Objects/Chest.h"
-#include "Engine/Objects/ItemTable.h"
 #include "Engine/Objects/SpriteObject.h"
 #include "Engine/OurMath.h"
 #include "Engine/Party.h"
 #include "Engine/Serialization/CompositeImages.h"
 #include "Engine/SpellFxRenderer.h"
+#include "Engine/Tables/ItemTable.h"
 #include "Engine/Tables/TileFrameTable.h"
 #include "Engine/Time.h"
 #include "Engine/TurnEngine/TurnEngine.h"
