+++ src/Engine/Objects/Actor.cpp
@@ -16,7 +16,6 @@
 #include "Engine/Graphics/Vis.h"
 #include "Engine/Localization.h"
 #include "Engine/LOD.h"
-#include "Engine/Objects/ItemTable.h"
 #include "Engine/Objects/ObjectList.h"
 #include "Engine/Objects/SpriteObject.h"
 #include "Engine/OurMath.h"
@@ -25,6 +24,7 @@
 #include "Engine/SpellFxRenderer.h"
 #include "Engine/Spells/CastSpellInfo.h"
 #include "Engine/AttackList.h"
+#include "Engine/Tables/ItemTable.h"
 #include "Engine/Tables/FactionTable.h"
 #include "Engine/Tables/AwardTable.h"
 #include "Engine/Time.h"
