+++ src/Engine/Objects/Chest.cpp
@@ -11,9 +11,9 @@
 #include "Engine/MapInfo.h"
 #include "Engine/Objects/Actor.h"
 #include "Engine/Objects/Items.h"
-#include "Engine/Objects/ItemTable.h"
 #include "Engine/Objects/ObjectList.h"
 #include "Engine/Objects/SpriteObject.h"
+#include "Engine/Tables/ItemTable.h"
 #include "Engine/Serialization/LegacyImages.h"
 #include "Engine/OurMath.h"
 #include "Engine/Party.h"
