+++ src/Engine/Objects/Items.cpp
@@ -8,9 +8,9 @@
 #include "Engine/Engine.h"
 #include "Engine/Localization.h"
 #include "Engine/Objects/ItemEnchantment.h"
-#include "Engine/Objects/ItemTable.h"
 #include "Engine/Objects/Monsters.h"
 #include "Engine/Objects/PlayerEnums.h"
+#include "Engine/Tables/ItemTable.h"
 #include "Engine/OurMath.h"
 #include "Engine/Party.h"
