+++ src/Engine/Spells/Spells.cpp
@@ -29,6 +29,8 @@ std::array<TownPortalData, 6> TownPortalList =  // 4ECBB8
       {Vec3i(-158, 7624, 1), 512, 0, 7, 0},     // celeste
       {Vec3i(-1837, -4247, 65), 65, 0, 8, 0}}};  // pit
 
+std::string pSpellsTXT_Raw;
+
 struct SpellStats *pSpellStats = nullptr;
 
 /**
