+++ src/GUI/UI/UICharacter.cpp
@@ -12,10 +12,10 @@
 #include "Engine/Graphics/Viewport.h"
 #include "Engine/LOD.h"
 #include "Engine/Localization.h"
-#include "Engine/Objects/ItemTable.h"
 #include "Engine/Party.h"
 #include "Engine/Spells/CastSpellInfo.h"
 #include "Engine/Time.h"
+#include "Engine/Tables/ItemTable.h"
 #include "Engine/Tables/AwardTable.h"
 
 #include "GUI/GUIWindow.h"
