+++ src/GUI/UI/UIGuilds.cpp
@@ -9,7 +9,7 @@
 #include "Engine/LOD.h"
 #include "Engine/Localization.h"
 #include "Engine/Objects/Items.h"
-#include "Engine/Objects/ItemTable.h"
+#include "Engine/Tables/ItemTable.h"
 #include "Engine/Party.h"
 #include "Engine/PriceCalculator.h"
