+++ src/GUI/UI/UIPartyCreation.cpp
@@ -10,8 +10,8 @@
 #include "Engine/Graphics/Viewport.h"
 #include "Engine/LOD.h"
 #include "Engine/Localization.h"
-#include "Engine/Objects/ItemTable.h"
 #include "Engine/Party.h"
+#include "Engine/Tables/ItemTable.h"
 #include "Engine/Tables/IconFrameTable.h"
 #include "Engine/Time.h"
