+++ src/Io/Mouse.cpp
@@ -9,7 +9,7 @@
 #include "Engine/Graphics/Vis.h"
 #include "Engine/LOD.h"
 #include "Engine/Objects/Actor.h"
-#include "Engine/Objects/ItemTable.h"
+#include "Engine/Tables/ItemTable.h"
 #include "Engine/Party.h"
 #include "Engine/TurnEngine/TurnEngine.h"
